[PIPE-374] Factoring out the pyspark dependent modules so only imported when a spark-based mgmt command is called
